Support empty files in --parameters list.
When using `stackup` in a Makefile which has automatically generated
targets, i would like to be able to deploy a stack which requires no
parameters using the same command invocation as a stack which does
require parameters.  To achieve this, i sometimes want to pass an
empty file in to the `--parameters` option of stackup.

This patch initialises the input to an empty hash if it doesn't have a
value (as is the case when reading an empty file) and adds a spec
around that behaviour.

* lib/stackup/parameters.rb (new): Initialise the `arg` variable if it
  is false (this would otherwise cause `#hashify` to crash).
* spec/stackup/parameters_spec.rb: Add a spec to check behaviour of
  `Parameters#new` when given `false`.