Make CFn diff context size configurable.
As a user, i sometimes want to be able to see *less* context around my
CloudFormation diffs than the entire file, because for small changes,
the full file can be an overwhelming screen-full.

This patch doesn't change the default behaviour of Stackup, it keeps
the default of 10,000 lines of context as defined in the Diffy
library.