Convert param input to String
The ruby AWS SDK always expects string input. Even if your
cloudformation parameter type is `Number`, it still wants it to be a
ruby String. The validation of parameter value will happen in
cloudformation land.

Currently stackup is just passing through the raw YAML value to the SDK
as is - which means it is up to the caller to understand the nuances on
the ruby AWS SDK (where everything must be a string). Given stackup is
supposed to "help" reduce cognitive load, I think it's ok for it to help
convert our YAML values in this instance.